+++ mp_flipper_modflipperzero.h
@@ -244,5 +244,5 @@ bool mp_flipper_pwm_is_running(uint8_t raw_pin);
 
 #define MP_FLIPPER_INFRARED_RX_DEFAULT_TIMEOUT (1000000)
 
-uint32_t* mp_flipper_infrared_receive(uint32_t timeout, size_t* length);
-bool mp_flipper_infrared_transmit(uint32_t* signal, size_t length);
+uint16_t* mp_flipper_infrared_receive(uint32_t timeout, size_t* length);
+bool mp_flipper_infrared_transmit(uint16_t* signal, size_t length);
